plantch/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.2
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.0
wcwidth==0.2.5
